Add wrapper around list view table component
